virdevmapper: Don't error on kernels without DM support
In one of my latest patch (v6.6.0~30) I was trying to remove
libdevmapper use in favor of our own implementation. However, the
code did not take into account that device mapper can be not
compiled into the kernel (e.g. be a separate module that's not
loaded) in which case /proc/devices won't have the device-mapper
major number and thus virDevMapperGetTargets() and/or
virIsDevMapperDevice() fails.

Fixes: 22494556542c676d1b9e7f1c1f2ea13ac17e1e3e
Reported-by: Andrea Bolognani <abologna@redhat.com>
Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Message-Id: <8ea97d64d41f7eed6309b26e9291554cd114a614.1597674163.git.mprivozn@redhat.com>
Tested-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>