virdevmapper: Deal with unloading dm module
The following situation can happen: we've initialized DM major
successfully. But later, the devmapper module(s) was removed and
thus kernel lost its ability to handle multipath devices. More
importantly, the /dev/mapper/control file is removed and thus our
attempt to open it will fail. Note, we will attempt to open it
only after we've found devmapper major number successfully in one
of the previous runs. Anyway, if it so happens that the module
was unloaded then we have to reset the major number we remembered
from one of the previous runs and not report error.

Fixes: 22494556542c676d1b9e7f1c1f2ea13ac17e1e3e
Reported-by: Andrea Bolognani <abologna@redhat.com>
Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Message-Id: <f028a89a12e02265dc2fca4690df7ec55727d6bf.1597674163.git.mprivozn@redhat.com>
Tested-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>