qemu: don't needlessly unset close callback during perform phase
During API call connection is referenced and close callback is called when
connection is disposed. Thus during API call close callback cannot be triggered
and we don't need to unset callback on API call duration.

This code was added in [1] and commit does not explain this part of the
patch.

[1] 1fdc53c3: qemu: Avoid dangling migration-out job when client dies

Signed-off-by: Nikolay Shirokovskiy <nshirokovskiy@virtuozzo.com>
Message-Id: <1597746396-983813-3-git-send-email-nshirokovskiy@virtuozzo.com>
Reviewed-by: Michal Privoznik <mprivozn@redhat.com>