meson/src: add missing augeas tests
Most of our augeas files are generated during meson setup into build
directory and we were running augeas tests only for these files.

However, we have some other augeas and config files that are not
modified during meson setup and they are only in source directories.
In order to run tests for these files we need to provide different path
to both source and build directories.

Reported-by: Michal Privoznik <mprivozn@redhat.com>
Signed-off-by: Pavel Hrdina <phrdina@redhat.com>
Message-Id: <f1601db81473667abbaed2e15f4779d5f9834932.1597758182.git.phrdina@redhat.com>
Reviewed-by: Michal Privoznik <mprivozn@redhat.com>