meson: Introduce rpath option
Right now we're unconditionally adding RPATH information to the
installed binaries and libraries, but that's not always desired.

Debian explicitly passes --disable-rpath to configure, and while
I haven't been able to find the same option in the spec file for
either Fedora or RHEL, by running

  $ readelf -d /usr/bin/virsh | grep PATH

I can see that the information is not present, so I assume they
also strip it somehow.

Both Debian and Fedora have wiki pages encouraging packagers to
avoid setting RPATH:

  https://wiki.debian.org/RpathIssue
  https://fedoraproject.org/wiki/RPath_Packaging_Draft

Given the above I'm not actually sure whether there even is a
valid usecase for RPATH, but I will openly admit I don't
understand the problem space well enough to pass judgement. So,
assuming there are scenarios where we want RPATH information to
be present, our only course of action is making its inclusion
configurable, just like it was with autotools.

This commit introduces the 'rpath' option, which is enabled by
default and can be turned off to avoid including RPATH
information in the distributed binaries and libraries.

The implementation is pretty naive and arguably verging on
unmaintainable, but unfortunately I lack the Meson-fu necessary
to do any better than this. Thankfully we have developers on
the list who are well versed in the build system and will
certainly come up with a reasonable solution :)

This commit is better viewed with 'git show -w'.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>
Message-Id: <20200819101808.144869-2-abologna@redhat.com>