qemu_domainjob: add `saveDomainStatus` as a callback function to jobs
The function `qemuDomainObjSaveStatus` required an access to
`virQEMUDriverPtr`.
To make jobs hypervisor-agnostic we remove this funciton
and replace it with a callback function from `qemuDomainJob`

Removal of `virQEMUDriverPtr` as parameter resulted in
removal of the same from function, where it was pass.
All of such references were removed as the variable
was no longer required.

Signed-off-by: Prathamesh Chavan <pc44800@gmail.com>
Message-Id: <20200823203808.97348-8-pc44800@gmail.com>