README: Drop symlink
Having a README file called "README" is necessary when using
autotools, and for quite some time now we've been complying with
that requirement by having a symlink with that name pointing to
README.rst, where the actual contents live. Now that we've moved
to Meson, we can drop the symlink.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>
Message-Id: <20200823221615.27750-3-abologna@redhat.com>
Reviewed-by: Erik Skultety <eskultet@redhat.com>