meson: link libm
****************
(So here is a separate patch to add linking of libm. Care to ACK it?
Also, what are peoples' opinions of pushing these patches now, so that
they'll be in the upcoming release? I've put them on a private gitlab
branch so that the CI is run (and found two mingw build problems :-)),
except I haven't been able to make the cirrus-ci thing that builds
freebsd and macos to work.)
****************

On some platforms libm (needed for the pow() function) isn't being
linked in somehow. This patch adds the necessary bits to assure that
it's linked in when necessary.

Suggested-by: Michal Privoznik <mprivozn@redhat.com>
Signed-off-by: Laine Stump <laine@redhat.com>
Message-Id: <20200826203510.211592-1-laine@redhat.com>
Reviewed-by: Daniel P. Berrangé <berrange@redhat.com>