wireshark: Don't include config.h
While both Debian and Fedora include the header in their development
packages for Wireshark, that's not something that the upstream
developers intended and arguably quite wrong, as config.h is obviously
intended to only be used to drive the compilation of Wireshark itself.
The Arch Linux package behaves like the upstream Wireshark package, and
thus libvirt fails to build there.

It seems that there are multiple bugs to be addressed:

  * libvirt shouldn't include config.h;

  * Debian and Fedora shouldn't be shipping config.h in their Wireshark
    packages;

  * Wireshark should not use config.h defines such as HAVE_PLUGINS in
    its public headers, and define a public variant of them instead.

This patch takes care of the first one.

https://gitlab.com/libvirt/libvirt/-/issues/74
Signed-off-by: Andrea Bolognani <abologna@redhat.com>
Message-Id: <20200903104329.35960-1-abologna@redhat.com>
Reviewed-by: Laine Stump <laine@redhat.com>